change clientList in a single component
